Clean imports and Remove signup box
Add admin sign up page

The difference between the admin sign up page and the default sign
up page is that (1) there is a textfield displaying the user is an
admin and (2) there will be a verification for the user to be there.
If the user is not verified, the user will be redirected to the
regularly sign up page. (3) an email will be filled in for them