Update schema to include difficulty and language in session
Remove status from session

Update endpoints to take language and difficulty

Create session endpoint takes a roomid as well to identify the
session, since the call is made from matching-service. Username
is used as it is unique and used to identify the user in
matching-service.

Update frontend api calls to match data available in profilepage