Create responses for users
When users click next question, value of the code editor is
retrieved and a update/insert request is sent to the backend. This
will record the current response for the question, and it is shared
between both users.

Add mobx for easier state handling across components

Minor updates to Ui for Question Panel

Add question records to history table

Previous sessions and questions are retrieved from database when
page is loaded. Questions are matched by id to responses. The
session history is mapped and shown under the history card, and
response for each question is loaded as well.